add precheck for delete icon function

+++ databases/sql/tracker/040-icons.sql
@@ -46,10 +46,29 @@ $$
 declare
     count integer;
 begin
-    -- an icon cannot be deleted if it is in use (project, workflow, ticketType, team, etc)
-    delete from icons where id = entityId;
-    get diagnostics count = ROW_COUNT;
-    return count;
+    if deleteIconPreCheck(entityId) then
+        delete from icons where id = entityId;
+        get diagnostics count = ROW_COUNT;
+        return count;
+    else
+        return 0; --error state
+    end if;
+end;
+$$ language plpgsql;
+/*
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+ * deleteIconPreCheck() function
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+ */
+create or replace function deleteIconPreCheck(entityId uuid) returns boolean as
+$$
+begin
+    /*
+     * This is currently a placeholder.  This function should be overloaded later
+     * as record consumers are defined so that we can prevent any record here from being
+     * deleted until all related records are removed.
+     */
+    return true;
 end;
 $$ language plpgsql;
 /*
